backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.2
mistralai==0.1.2
groq==0.4.2
httpx==0.25.2
pydantic==2.5.0

